refactor lobby page rel: #75
