+++ apps/game/src/components/lobby/RoomCode.tsx
@@ -3,7 +3,7 @@ type RoomCodeProps = {
 };
 
 export const RoomCode = ({ code }: RoomCodeProps) => (
-  <div className="font-raleway rounded-lg bg-slate-700 py-3 text-center text-4xl font-bold">
+  <div className="font-raleway w-full max-w-xs rounded-lg bg-slate-700 py-4 text-center text-4xl font-bold">
     {code}
   </div>
 );
